refactor: extract sidebar profile section
