+++ web/src/app/layout.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 
 import './globals.css';
-import { AppSidebar } from '@/components/AppSidebar';
+import { AppSidebar } from '@/components/nav/AppSidebar';
 import NavigationBar from '@/components/NavigationBar';
 import { SidebarTrigger, SidebarInset } from '@/components/ui/sidebar';
 import { Toaster } from '@/components/ui/sonner';
